first version of choosing proposal-development method (#19)
* removed (again) VS project files

* move changes to proper branch (#17)

* ignore working files of Visual Studio

* Prakash0 (#15)

* Scenario Description changed from all the tables

* work in porgress for bargainVote

* lost changes

* Scenario Id generation changed

* comments absorbed

* Work in progress

* BargainCoord changes

* merging with prakash0

* SqlTable Changes merged

* deleted the VS project files

* Testing (#16)

* Prakash0 (#15)

* Scenario Description changed from all the tables

* work in porgress for bargainVote

* lost changes

* Scenario Id generation changed

* comments absorbed

* Work in progress

* BargainCoord changes

* merging with prakash0

* SqlTable Changes merged

* deleted the VS project files

* add comments relating to ideals

* implemented first three models of SMP bargaining

* choose from 3 very basic options for making proposed bargains

* modify PRNG

* note crypto stuff

+++ KTAB/kmodel/libsrc/kmodel.cpp
@@ -65,7 +65,7 @@ Model::Model(PRNG * r, string desc) {
     if (0 == desc.length()) {
         auto utcBuff = newChars(200);
 		
-        std::strftime(utcBuff, 150, "Scenario-UTC-%Y-%m-%d-%H%M-%S", gmtime(&start_time));
+        std::strftime(utcBuff, 150, "Scenario-UTC-%Y-%m-%u-%H%M-%S", gmtime(&start_time));
         cout << "No scenario description provided to Model::Model, " << endl;
         cout << "generating default name from UTC start time." << endl << flush;
         scenName = utcBuff;
@@ -83,7 +83,7 @@ Model::Model(PRNG * r, string desc) {
 		sprintf(utcBuffId, "%s_%u", desc, milliseconds);
 	}
 	//get the hash
-	unsigned int scenIdhash = (std::hash < std::string> () (utcBuffId))   ;
+	uint64_t scenIdhash = (std::hash < std::string> () (utcBuffId))   ;
 	sprintf(hshCode, "0x%032llX", scenIdhash);
 	scenId = hshCode;
 	delete hshCode;
